requirements.txt
PyYAML==3.13
netifaces==0.10.8
flask==1.0.2
flask_restful==0.3.7
pykube==0.15.0
coverage==4.5.1